Also query all external exchanges for historical accounting
This commit moves some querying trades + asset movements code from
rotkehlchen super object to the EventsHistorian
