+++ rotkehlchen/history/events.py
@@ -1,17 +1,27 @@
 import logging
+from collections import defaultdict
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, List, Optional, Tuple, Union, cast
+from typing import TYPE_CHECKING, Any, Dict, List, Optional, Tuple, Union, cast, overload
+
+from typing_extensions import Literal
 
 from rotkehlchen.chain.ethereum.trades import AMMTRADE_LOCATION_NAMES, AMMTrade, AMMTradeLocations
 from rotkehlchen.constants.misc import ZERO
 from rotkehlchen.db.ledger_actions import DBLedgerActions
 from rotkehlchen.errors import RemoteError
 from rotkehlchen.exchanges.data_structures import AssetMovement, Loan, MarginPosition, Trade
-from rotkehlchen.exchanges.manager import ExchangeManager
+from rotkehlchen.exchanges.exchange import ExchangeInterface
+from rotkehlchen.exchanges.manager import ALL_SUPPORTED_EXCHANGES, ExchangeManager
 from rotkehlchen.exchanges.poloniex import process_polo_loans
 from rotkehlchen.fval import FVal
 from rotkehlchen.logging import RotkehlchenLogsAdapter
-from rotkehlchen.typing import EthereumTransaction, Location, Timestamp
+from rotkehlchen.typing import (
+    EXTERNAL_EXCHANGES,
+    EXTERNAL_LOCATION,
+    EthereumTransaction,
+    Location,
+    Timestamp,
+)
 from rotkehlchen.user_messages import MessagesAggregator
 from rotkehlchen.utils.accounting import action_get_timestamp
 from rotkehlchen.utils.misc import timestamp_to_date
@@ -29,7 +39,8 @@ log = RotkehlchenLogsAdapter(logger)
 # Number of steps excluding the connected exchanges. Current query steps:
 # eth transactions
 # ledger actions
-# external trades: balancer, uniswap
+# external location trades -> len(EXTERNAL_LOCATION)
+# amm trades len(AMMTradeLocations)
 # makerdao dsr
 # makerdao vaults
 # yearn vaults
@@ -38,7 +49,7 @@ log = RotkehlchenLogsAdapter(logger)
 # aave lending
 # eth2
 # Please, update this number each time a history query step is either added or removed
-NUM_HISTORY_QUERY_STEPS_EXCL_EXCHANGES = 11
+NUM_HISTORY_QUERY_STEPS_EXCL_EXCHANGES = 9 + len(EXTERNAL_LOCATION) + len(AMMTradeLocations)
 FREE_LEDGER_ACTIONS_LIMIT = 50
 
 HistoryResult = Tuple[
@@ -50,6 +61,16 @@ HistoryResult = Tuple[
     List['DefiEvent'],
     List['LedgerAction'],
 ]
+TRADES_LIST = List[Union[Trade, AMMTrade]]
+
+FREE_TRADES_LIMIT = 250
+FREE_ASSET_MOVEMENTS_LIMIT = 100
+
+LIMITS_MAPPING = {
+    'trade': FREE_TRADES_LIMIT,
+    'asset_movement': FREE_ASSET_MOVEMENTS_LIMIT,
+    'ledger_action': FREE_LEDGER_ACTIONS_LIMIT,
+}
 
 
 def limit_trade_list_to_period(
@@ -103,6 +124,11 @@ class EventsHistorian():
         )
 
     def _reset_variables(self) -> None:
+        # Keeps how many trades we have found per location. Used for free user limiting
+        self.actions_per_location: Dict[str, Dict[Location, int]] = {
+            'trade': defaultdict(int),
+            'asset_movement': defaultdict(int),
+        }
         self.processing_state_name = 'Starting query of historical events'
         self.progress = ZERO
         db_settings = self.db.get_settings()
@@ -113,6 +139,56 @@ class EventsHistorian():
         step += 1
         self.progress = FVal(step / total_steps) * 100
         return step
+
+    @overload
+    def _apply_actions_limit(
+            self,
+            location: Location,
+            action_type: Literal['trade'],
+            location_actions: TRADES_LIST,
+            all_actions: TRADES_LIST,
+    ) -> TRADES_LIST:
+        ...
+
+    @overload
+    def _apply_actions_limit(
+            self,
+            location: Location,
+            action_type: Literal['asset_movement'],
+            location_actions: List[AssetMovement],
+            all_actions: List[AssetMovement],
+    ) -> List[AssetMovement]:
+        ...
+
+    def _apply_actions_limit(
+            self,
+            location: Location,
+            action_type: Literal['trade', 'asset_movement'],
+            location_actions: Union[TRADES_LIST, List[AssetMovement]],
+            all_actions: Union[TRADES_LIST, List[AssetMovement]],
+    ) -> Union[TRADES_LIST, List[AssetMovement]]:
+        """Take as many actions from location actions and add them to all actions as the limit permits
+
+        Returns the modified (or not) all_actions
+        """
+        # If we are already at or above the limit return current actions disregarding this location
+        actions_mapping = self.actions_per_location[action_type]
+        current_num_actions = sum(x for _, x in actions_mapping.items())
+        limit = LIMITS_MAPPING[action_type]
+        if current_num_actions >= limit:
+            return all_actions
+
+        # Find out how many more actions can we return, and depending on that get
+        # the number of actions from the location actions and add them to the total
+        remaining_num_actions = limit - current_num_actions
+        if remaining_num_actions < 0:
+            remaining_num_actions = 0
+
+        num_actions_to_take = min(len(location_actions), remaining_num_actions)
+
+        actions_mapping[location] = num_actions_to_take
+        all_actions.extend(location_actions[0:num_actions_to_take])  # type: ignore
+        return all_actions
 
     def query_ledger_actions(
             self,
@@ -135,6 +211,265 @@ class EventsHistorian():
             actions = actions[:FREE_LEDGER_ACTIONS_LIMIT]
 
         return actions, original_length
+
+    def query_trades(
+            self,
+            from_ts: Timestamp,
+            to_ts: Timestamp,
+            location: Optional[Location],
+            only_cache: bool,
+    ) -> TRADES_LIST:
+        """Queries trades for the given location and time range.
+        If no location is given then all external, all exchange and DEX trades are queried.
+
+        If only_cache is given then only trades cached in the DB are returned.
+        No service is queried.
+
+        DEX Trades are queried only if the user has premium
+        If the user does not have premium then a trade limit is applied.
+
+        May raise:
+        - RemoteError: If there are problems connecting to any of the remote exchanges
+        """
+        trades: TRADES_LIST
+        if location is not None:
+            # clear the trades queried for this location
+            self.actions_per_location['trade'][location] = 0
+            trades = self.query_location_trades(from_ts, to_ts, location, only_cache)
+        else:
+            for given_location in ALL_SUPPORTED_EXCHANGES + [Location.EXTERNAL]:
+                # clear the trades queried for this location
+                self.actions_per_location['trade'][given_location] = 0
+            trades = self.query_location_trades(from_ts, to_ts, Location.EXTERNAL, only_cache)
+            # Look for trades that might be imported from CSV files
+            for csv_location in EXTERNAL_EXCHANGES:
+                trades.extend(self.query_location_trades(
+                    from_ts=from_ts,
+                    to_ts=to_ts,
+                    location=csv_location,
+                    only_cache=only_cache,
+                ))
+
+            for exchange in self.exchange_manager.iterate_exchanges():
+                all_set = {x.identifier for x in trades}
+                exchange_trades = exchange.query_trade_history(
+                    start_ts=from_ts,
+                    end_ts=to_ts,
+                    only_cache=only_cache,
+                )
+                # TODO: Really dirty. Figure out a better way.
+                # Since some of the trades may already be in the DB if multiple
+                # keys are used for a single exchange.
+                exchange_trades = [x for x in exchange_trades if x.identifier not in all_set]
+                if self.chain_manager.premium is None:
+                    trades = self._apply_actions_limit(
+                        location=exchange.location,
+                        action_type='trade',
+                        location_actions=exchange_trades,
+                        all_actions=trades,
+                    )
+                else:
+                    trades.extend(exchange_trades)
+
+            # for all trades we also need the trades from the amm protocols
+            if self.chain_manager.premium is not None:
+                for amm_location in AMMTradeLocations:
+                    amm_module_name = cast(AMMTRADE_LOCATION_NAMES, str(amm_location))
+                    amm_module = self.chain_manager.get_module(amm_module_name)
+                    if amm_module is not None:
+                        trades.extend(
+                            amm_module.get_trades(
+                                addresses=self.chain_manager.queried_addresses_for_module(amm_module_name),  # noqa: E501
+                                from_timestamp=from_ts,
+                                to_timestamp=to_ts,
+                                only_cache=only_cache,
+                            ),
+                        )
+        # return trades with most recent first
+        trades.sort(key=lambda x: x.timestamp, reverse=True)
+        return trades
+
+    def query_location_trades(
+            self,
+            from_ts: Timestamp,
+            to_ts: Timestamp,
+            location: Location,
+            only_cache: bool,
+    ) -> TRADES_LIST:
+        location_trades: TRADES_LIST
+        if location in EXTERNAL_LOCATION:
+            location_trades = self.db.get_trades(  # type: ignore  # list invariance
+                from_ts=from_ts,
+                to_ts=to_ts,
+                location=location,
+            )
+        elif location in AMMTradeLocations:
+            if self.chain_manager.premium is not None:
+                amm_module_name = cast(AMMTRADE_LOCATION_NAMES, str(location))
+                amm_module = self.chain_manager.get_module(amm_module_name)
+                if amm_module is not None:
+                    location_trades = amm_module.get_trades(  # type: ignore  # list invariance
+                        addresses=self.chain_manager.queried_addresses_for_module(amm_module_name),
+                        from_timestamp=from_ts,
+                        to_timestamp=to_ts,
+                        only_cache=only_cache,
+                    )
+        else:
+            # should only be an exchange
+            exchanges_list = self.exchange_manager.connected_exchanges.get(location)
+            if exchanges_list is None:
+                log.warning(
+                    f'Tried to query trades from {str(location)} which is either not an '
+                    f'exchange or not an exchange the user has connected to',
+                )
+                return []
+
+            location_trades = []
+            for exchange in exchanges_list:
+                all_set = {x.identifier for x in location_trades}
+                new_trades = exchange.query_trade_history(
+                    start_ts=from_ts,
+                    end_ts=to_ts,
+                    only_cache=only_cache,
+                )
+                # TODO: Really dirty. Figure out a better way.
+                # Since some of the trades may already be in the DB if multiple
+                # keys are used for a single exchange.
+                new_trades = [x for x in new_trades if x.identifier not in all_set]
+                location_trades.extend(new_trades)
+
+        trades: TRADES_LIST = []
+        if self.chain_manager.premium is None:
+            trades = self._apply_actions_limit(
+                location=location,
+                action_type='trade',
+                location_actions=location_trades,
+                all_actions=trades,
+            )
+        else:
+            trades = location_trades
+
+        return trades
+
+    def _query_and_populate_exchange_asset_movements(
+            self,
+            from_ts: Timestamp,
+            to_ts: Timestamp,
+            all_movements: List[AssetMovement],
+            exchange: Union[ExchangeInterface, Location],
+            only_cache: bool,
+    ) -> List[AssetMovement]:
+        """Queries exchange for asset movements and adds it to all_movements"""
+        all_set = {x.identifier for x in all_movements}
+        if isinstance(exchange, ExchangeInterface):
+            location = exchange.location
+            location_movements = exchange.query_deposits_withdrawals(
+                start_ts=from_ts,
+                end_ts=to_ts,
+                only_cache=only_cache,
+            )
+            # TODO: Really dirty. Figure out a better way.
+            # Since some of the asset movements may already be in the DB if multiple
+            # keys are used for a single exchange.
+            location_movements = [x for x in location_movements if x.identifier not in all_set]
+        else:
+            assert isinstance(exchange, Location), 'only a location should make it here'
+            assert exchange in EXTERNAL_EXCHANGES, 'only csv supported exchanges should get here'  # noqa : E501
+            location = exchange
+            # We might have no exchange information but CSV imported information
+            self.actions_per_location['asset_movement'][location] = 0
+            location_movements = self.db.get_asset_movements(
+                from_ts=from_ts,
+                to_ts=to_ts,
+                location=location,
+            )
+
+        movements: List[AssetMovement] = []
+        if self.chain_manager.premium is None:
+            movements = self._apply_actions_limit(
+                location=location,
+                action_type='asset_movement',
+                location_actions=location_movements,
+                all_actions=all_movements,
+            )
+        else:
+            all_movements.extend(location_movements)
+            movements = all_movements
+
+        return movements
+
+    def query_asset_movements(
+            self,
+            from_ts: Timestamp,
+            to_ts: Timestamp,
+            location: Optional[Location],
+            only_cache: bool,
+    ) -> List[AssetMovement]:
+        """Queries AssetMovements for the given location and time range.
+
+        If no location is given then all exchange asset movements are queried.
+        If only_cache is True then only what is already in the DB is returned.
+        If the user does not have premium then a limit is applied.
+        May raise:
+        - RemoteError: If there are problems connecting to any of the remote exchanges
+        """
+        movements: List[AssetMovement] = []
+        if location is not None:
+            # clear the asset movements queried for this exchange
+            self.actions_per_location['asset_movement'][location] = 0
+            if location in EXTERNAL_EXCHANGES:
+                movements = self._query_and_populate_exchange_asset_movements(
+                    from_ts=from_ts,
+                    to_ts=to_ts,
+                    all_movements=movements,
+                    exchange=location,
+                    only_cache=only_cache,
+                )
+            else:
+                exchanges_list = self.exchange_manager.connected_exchanges.get(location)
+                if exchanges_list is None:
+                    log.warning(
+                        f'Tried to query deposits/withdrawals from {str(location)} which is '
+                        f'either not an exchange or not an exchange the user has connected to',
+                    )
+                    return []
+
+                # clear the asset movements queried for this exchange
+                self.actions_per_location['asset_movement'][location] = 0
+                for exchange in exchanges_list:
+                    self._query_and_populate_exchange_asset_movements(
+                        from_ts=from_ts,
+                        to_ts=to_ts,
+                        all_movements=movements,
+                        exchange=exchange,
+                        only_cache=only_cache,
+                    )
+        else:
+            for exchange_location in ALL_SUPPORTED_EXCHANGES:
+                # clear the asset movements queried for this exchange
+                self.actions_per_location['asset_movement'][exchange_location] = 0
+            # we may have DB entries due to csv import from supported locations
+            for external_location in EXTERNAL_EXCHANGES:
+
+                movements = self._query_and_populate_exchange_asset_movements(
+                    from_ts=from_ts,
+                    to_ts=to_ts,
+                    all_movements=movements,
+                    exchange=external_location,
+                    only_cache=only_cache,
+                )
+            for exchange in self.exchange_manager.iterate_exchanges():
+                self._query_and_populate_exchange_asset_movements(
+                    from_ts=from_ts,
+                    to_ts=to_ts,
+                    all_movements=movements,
+                    exchange=exchange,
+                    only_cache=only_cache,
+                )
+
+        # return movements with most recent first
+        movements.sort(key=lambda x: x.timestamp, reverse=True)
+        return movements
 
     def get_history(
             self,
@@ -217,16 +552,18 @@ class EventsHistorian():
             empty_or_error += '\n' + msg
         step = self._increase_progress(step, total_steps)
 
-        # Include the external trades in the history
-        self.processing_state_name = 'Querying external trades history'
-        external_trades = self.db.get_trades(
-            # We need to have history of trades since before the range
-            from_ts=Timestamp(0),
-            to_ts=end_ts,
-            location=Location.EXTERNAL,
-        )
-        history.extend(external_trades)
-        step = self._increase_progress(step, total_steps)
+        # Include all external trades and trades from external exchanges
+        for location in EXTERNAL_LOCATION:
+            self.processing_state_name = f'Querying {location} trades history'
+            external_trades = self.query_location_trades(
+                # We need to have history of trades since before the range
+                from_ts=Timestamp(0),
+                to_ts=end_ts,
+                location=location,
+                only_cache=True,
+            )
+            history.extend(external_trades)
+            step = self._increase_progress(step, total_steps)
 
         # include the ledger actions
         self.processing_state_name = 'Querying ledger actions history'
@@ -311,7 +648,7 @@ class EventsHistorian():
                 to_timestamp=end_ts,
                 addresses=self.chain_manager.queried_addresses_for_module('aave'),
             ))
-        self._increase_progress(step, total_steps)
+        step = self._increase_progress(step, total_steps)
 
         # include eth2 staking events
         eth2 = self.chain_manager.get_module('eth2')
